TrustDaemon building model at periodic intervals
Model code not yet enabled, so mostly a stub for now.
Add a configuration for trust with an interval for daemon. Daemon is structured
same as Gossip daemon in terms of classes used. Relies on trustStorage data assumed to be updated by gossip and updates internal storage with predictions.

This is for feature parity with existing codebase. Long term requirements will need model updates.